gs.glsl improvements
Hey! Really liked your particle sim, tried it out myself! Here are my improvements to your geometry shader code: less normalizations and matrix - vector multiplications. I took out the incrementing of the gl_Position var, since it seems like it was left in. Seems to give the same output as the original code, but don't take my word for it :)